requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
altair==5.1.1
astroid==2.15.5
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
blinker==1.6.2
botocore==1.29.137
Brotli==1.0.9
cachetools==5.3.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
cycler==0.11.0
dash==2.5.1
dash-bootstrap-components==1.1.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.2
decorator==5.1.1
deta==1.2.0
dill==0.3.5.1
distlib==0.3.4
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.6.0
Flask==2.1.2
Flask-Compress==1.12
fonttools==4.34.4
frozenlist==1.3.1
gitdb==4.0.10
GitPython==3.1.37
greenlet==2.0.1
idna==3.4
importlib-metadata==4.12.0
ipykernel==6.15.1
ipython==8.4.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
jsonschema==4.7.2
jupyter-client==7.3.4
jupyter-core==4.11.1
kafka-python==2.0.2
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
loguru==0.6.0
lxml==4.9.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.2
multiprocess==0.70.13
mysql-connector-python==8.0.31
nbformat==5.4.0
nest-asyncio==1.5.5
nltk==3.7
numpy==1.22.4
packaging==21.3
pandas==1.4.3
parso==0.8.3
pathos==0.2.9
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
plotly==5.17.0
pox==0.3.1
ppft==1.7.6.5
prompt-toolkit==3.0.30
protobuf==4.24.3
psutil==5.9.1
psycopg2-binary==2.9.5
pure-eval==0.2.2
py4j==0.10.9.5
pyarrow==13.0.0
pycparser==2.21
pydeck==0.8.1b0
PyExecJS==1.5.1
Pygments==2.16.1
PyMySQL==1.0.2
pyparsing==3.0.9
pyrsistent==0.18.1
pyspark==3.3.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.1
pywin32==304
pyzmq==23.2.0
regex==2022.6.2
requests==2.28.1
rich==13.5.3
s3transfer==0.6.1
six==1.16.0
smmap==5.0.1
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
stack-data==0.3.0
streamlit==1.27.0
streamlit-option-menu==0.3.6
tenacity==8.2.3
textblob==0.17.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.8
toolz==0.12.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
translators==5.3.1
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.13
vaderSentiment==3.3.2
validators==0.22.0
virtualenv==20.14.1
watchdog==3.0.0
wcwidth==0.2.5
Werkzeug==2.1.2
win32-setctime==1.1.0
wordcloud==1.8.2.2
wrapt==1.15.0
yarl==1.8.1
zipp==3.8.0
